Split sleep tests and pure tests
